Use "gox" from local fork of "authelia/gox:master" repo
The "gox" source in the Hashicorp build script has been abandoned
and doesn't work on current ARM instances.  Someone forked that
repo and has updated it enough that it can be used to build a
working binary.  This change directs the "serf" build script to
reference a local copy of that fork.
a